+++ src/main/java/org/folio/linked/data/service/profile/PreferredProfileServiceImpl.java
@@ -10,10 +10,12 @@ import lombok.extern.log4j.Log4j2;
 import org.folio.linked.data.domain.dto.ProfileMetadata;
 import org.folio.linked.data.exception.RequestProcessingExceptionBuilder;
 import org.folio.linked.data.model.entity.PreferredProfile;
+import org.folio.linked.data.model.entity.ResourceTypeEntity;
 import org.folio.linked.data.model.entity.pk.PreferredProfilePk;
 import org.folio.linked.data.repo.PreferredProfileRepository;
 import org.folio.linked.data.repo.ProfileRepository;
 import org.folio.linked.data.repo.ResourceTypeRepository;
+import org.folio.spring.FolioExecutionContext;
 import org.springframework.lang.Nullable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -28,21 +30,30 @@ public class PreferredProfileServiceImpl implements PreferredProfileService {
   private final RequestProcessingExceptionBuilder exceptionBuilder;
   private final ResourceTypeRepository typeRepository;
   private final ProfileRepository profileRepository;
+  private final FolioExecutionContext executionContext;
 
   @Override
-  public void setPreferredProfile(UUID userId, Integer profileId, String resourceTypeUri) {
+  public void setPreferredProfile(Integer profileId, String resourceTypeUri) {
     var profile =  profileRepository.findById(profileId)
       .orElseThrow(() -> exceptionBuilder.notFoundLdResourceByIdException("Profile", String.valueOf(profileId)));
-    var resourceType = typeRepository.findByUri(resourceTypeUri);
+    var resourceType = getResourceTypeByUri(resourceTypeUri);
     var preferredProfile = new PreferredProfile()
-      .setId(new PreferredProfilePk(userId, resourceType.getHash()))
+      .setId(new PreferredProfilePk(executionContext.getUserId(), resourceType.getHash()))
       .setProfile(profile);
     preferredProfileRepository.save(preferredProfile);
   }
 
   @Override
+  public void deletePreferredProfile(String resourceTypeUri) {
+    var resourceType = getResourceTypeByUri(resourceTypeUri);
+    var idToDelete = new PreferredProfilePk(executionContext.getUserId(), resourceType.getHash());
+    preferredProfileRepository.deleteById(idToDelete);
+  }
+
+  @Override
   @Transactional(readOnly = true)
-  public List<ProfileMetadata> getPreferredProfiles(UUID userId, @Nullable String resourceTypeUri) {
+  public List<ProfileMetadata> getPreferredProfiles(@Nullable String resourceTypeUri) {
+    var userId = executionContext.getUserId();
     var preferredProfiles = ofNullable(resourceTypeUri)
       .map(uri -> getPreferredProfile(userId, uri).stream().toList())
       .orElseGet(() -> getPreferredProfiles(userId));
@@ -59,7 +70,12 @@ public class PreferredProfileServiceImpl implements PreferredProfileService {
   }
 
   private Optional<PreferredProfile> getPreferredProfile(UUID userId, String resourceTypeUri) {
-    var resourceTypeId = typeRepository.findByUri(resourceTypeUri).getHash();
+    var resourceTypeId = getResourceTypeByUri(resourceTypeUri).getHash();
     return preferredProfileRepository.findById(new PreferredProfilePk(userId, resourceTypeId));
+  }
+
+  private ResourceTypeEntity getResourceTypeByUri(String resourceTypeUri) {
+    return ofNullable(typeRepository.findByUri(resourceTypeUri))
+      .orElseThrow(() -> exceptionBuilder.badRequestException("Invalid Resource Type in request", resourceTypeUri));
   }
 }