+++ NEWS.md
@@ -26,6 +26,11 @@
 - Update to use extent object property exclusively, removing extent literal property. Update profiles to match. [MODLD-388](https://folio-org.atlassian.net/browse/MODLD-388)
 - Restore method `ResourceGraphService -> breakEdgesAndDelete` back to the codebase [MODLD-790](https://folio-org.atlassian.net/browse/MODLD-790)
 - Add validation for series title [MODLD-786](https://folio-org.atlassian.net/browse/MODLD-786)
+- New API to delete preferred profile for a user [MODLD-795](https://folio-org.atlassian.net/browse/MODLD-795)
+- Update API to support DATES_OF_PUBLICATION_NOTE [MODLD-808](https://folio-org.atlassian.net/browse/MODLD-808)
+- Enhance API to support characteristic edge of Work [MODLD-800](https://folio-org.atlassian.net/browse/MODLD-800)
+- Update profiles to support book format handling [MODLD-726](https://folio-org.atlassian.net/browse/MODLD-726)
+- Enhance `GET import-supported` API to check if MARC record is a Serial [MODLD-809](https://folio-org.atlassian.net/browse/MODLD-809)
 - Update profiles to support revised language handling [MODLD-362](https://folio-org.atlassian.net/browse/MODLD-362)
 
 ## 1.0.4 (04-24-2025)